[LF-12544][SRTM]re-init PDM DMA if some periods were missed
It has been observed that SRTM task can be blocked busy waiting
for answers from the kernel, and in case of heavy kernel load this
wait can be longer than a period.
In that case the procedure in charge of triggering the DMA is not
scheduled soon enough, causing the whole process to block indefinitely.
This patch detects this case and re-init the local buffer and the DMA to
restart the PDM data transfers in a clean way.