[MCUX-72490] Add "WIFI_BOARD_<board>" macro definition with Kconfig
[Description]
Currently when compiling the wifi_nxp with sdk next gen, it is still required
to manually define a macro for the wifi board, otherwise it can't build.

[Root Cause]
The definition of the "WIFI_BOARD_<>" macro be handled with Kconfig within the SDK.

[Fix]
Add MCUX_COMPONENT_component.wifi_bt_module.board_rw61x Kconfig.

[Testing]
Pass

Signed-off-by: Gang Li <gang.li_1@nxp.com>