[MCUX-74717][mimxrt700evk] Fail to start RingTone in handsfree
When there is a incoming call from peer device, mimxrt700evk has no
response.
It hanged in SAI0_DriverIRQHandler() since the FEF bit of SAI0->TCSR has
never been cleared.

In general, when using fsl_adapter_sai.c, the status bit clearing is
handled by the adapter.
Please noted that the issue may not occur in all cases, which is related
to the timing of DMA data processing.

In the past, SAI started counting from SAI1, and mimxrt700evk is the
first device to start from SAI0. Therefore, there were no such issue
with the adapter before.
But this also means that as devices expand, adapter may continue to be
updated to be compatible with all different devices.

Signed-off-by: Jiawei Yang <jiawei.yang_1@nxp.com>