[MCUX-80426][OSA][threadx] Disabling/Enabling interrupt can't be nested.
Fix issue that the `OSA_InterruptDisable()` and `OSA_InterruptEnable()`
calls can't be nested in threadx.

Modify `OSA_DisableIRQGlobal` and `OSA_EnableIRQGlobal` in threadx to
disable/enable all interrupts using `DisableGlobalIRQ` and
`EnableGlobalIRQ` since `TX_DISABLE` and `TX_RESTORE` can be configured
to use BASEPRI to disable interrupt, which is not matched with this
API design.

Update comments to explain the API usage.

Signed-off-by: Can Wang <can.wang@nxp.com>