[MCUX-79380][str] Fix issue in StrFormatScanf.
The issue is that when starting with non-whitespace character, SCANF
cannot work. For example, `SCANF("test %x", &a);` When entering string
"test 0x11", `a` will be 0, not 0x11.

The root cause is that for non-whitespace character, the index is
increased unexpectedly.

Fix this issue by adding a condition that only when double `%`
occurs the index is increased.

Signed-off-by: Can Wang <can.wang@nxp.com>