[MCUX-80265][ELE_HSEB] Fix bug in ECC key import
During private key imports, if public key was given as NULL, the driver
computed the expected length of the public key and sent the request
with NULL pointer, but a nonzero public key length. This caused HSEB
to return an invalid parameter error.

Behavior is fixed by forcing a 0-value public key length if a NULL
public key is passed, so that the public part gets internally computed
by HSE_B.

Signed-off-by: Patrik Nemeth