[MCUX-85035] Fix issue that shell cannot receive data on FRDM-RW612.
The root cause is that after power-on `DbgConsole_Init` starts
receiving UART data while a byte is already present on the UART bus.
After that, the data is captured and `SerialManager_RxCallback` is
triggered. Because `openedReadHandleHead` is NULL,
`SerialManager_StartReading` in `SerialManager_RxCallback` doesn't
start reading next UART data. As a result, the UART hardware stops
receiving further data and the shell cannot get input.

Fix this issue by removing the `(NULL != readHandle)` check so that
`SerialManager_StartReading` in `SerialManager_RxCallback` can read
next UART data.

Signed-off-by: Can Wang <can.wang@nxp.com>